Refactor PersonalInfoForm to its own file
