+++ src/components/form/StepperFormActions.tsx
@@ -1,0 +1,37 @@
+import { Button } from "../ui/button"
+import { useStepper } from "../ui/stepper"
+
+export const StepperFormActions = () => {
+  const {
+    nextStep,
+    prevStep,
+    resetSteps,
+    hasCompletedAllSteps,
+    isLastStep,
+    isOptionalStep,
+    isDisabledStep,
+  } = useStepper()
+  return (
+    <div className="w-full flex justify-end gap-2">
+      {hasCompletedAllSteps ? (
+        <Button size="sm" onClick={resetSteps}>
+          Reset
+        </Button>
+      ) : (
+        <>
+          <Button
+            disabled={isDisabledStep}
+            onClick={prevStep}
+            size="sm"
+            variant="secondary"
+          >
+            Back
+          </Button>
+          <Button size="sm">
+            {isLastStep ? "Finish" : isOptionalStep ? "Skip" : "Next"}
+          </Button>
+        </>
+      )}
+    </div>
+  )
+}
